rename components ant titles 'Week' to 'Forecast'
