+++ src/App.jsx
@@ -4,7 +4,7 @@ import HeaderFC from './components/Header/Header'
 import MenuFC from './components/Menu/Menu'
 import { Layout } from 'antd'
 import CurrentContainer from './pages/Current/CurrentContainer'
-import WeekContainer from './pages/Week/WeekContainer'
+import ForecastContainer from './pages/Forecast/ForecastContainer'
 import DashboardContainer from './pages/Dashboard/DashboardContainer'
 import CalendarContainer from './pages/Calendar/CalendarContainer'
 import { useDispatch, useSelector } from 'react-redux'
@@ -43,7 +43,7 @@ const App = () => {
             <Routes>
               <Route path='/' element={<Navigate replace to='/current-weather' />} />
               <Route path='current-weather' element={<CurrentContainer />} />
-              <Route path='weather-forecast' element={<WeekContainer />} />
+              <Route path='weather-forecast' element={<ForecastContainer />} />
               <Route path='dashboard' element={<DashboardContainer />} />
               <Route path='calendar' element={<CalendarContainer />} />
             </Routes>
