+++ src/pages/Forecast/Forecast.jsx
@@ -1,18 +1,18 @@
 import { Container } from '@mui/material'
-import WeatherCard from './WeatherCard/WeatherCardS'
+import ForecastCard from './ForecastCard/ForecastCardS'
 import { useSelector } from 'react-redux'
 import { Space } from 'antd'
 import TitlePage from '../../components/common/Preloader/TitlePage/TitlePage'
-import s from './Week.module.css'
+import s from './Forecast.module.css'
 
-const Week = () => {
+const Forecast = () => {
   const forecastData = useSelector((state) => state.weather.forecastData)
   const dailyList = forecastData.list.filter((reading) => reading.dt_txt.includes('12:00:00')) // get data of every day per 12:00
   console.log(forecastData)
 
-  //cards days of week with data
+  // forecards cards with data
   const dailyFormatCards = () => {
-    return dailyList.map((day, index) => <WeatherCard day={day} key={index} />)
+    return dailyList.map((day, index) => <ForecastCard day={day} key={index} />)
   }
 
   return (
@@ -27,4 +27,4 @@ const Week = () => {
   )
 }
 
-export default Week
+export default Forecast
